Move COT commands to regulators
