+++ openbb_platform/extensions/regulators/integration/test_regulators_api.py
@@ -15,6 +15,9 @@ def headers():
     return {"Authorization": f"Basic {base64_bytes.decode('ascii')}"}
 
 
+# pylint: disable=redefined-outer-name
+
+
 @pytest.mark.parametrize(
     "params",
     [
@@ -153,6 +156,52 @@ def test_regulators_sec_sic_search(params, headers):
 
     query_str = get_querystring(params, [])
     url = f"http://0.0.0.0:8000/api/v1/regulators/sec/sic_search?{query_str}"
+    result = requests.get(url, headers=headers, timeout=10)
+    assert isinstance(result, requests.Response)
+    assert result.status_code == 200
+
+
+@pytest.mark.parametrize(
+    "params",
+    [
+        ({"query": "grain", "provider": "nasdaq"}),
+    ],
+)
+@pytest.mark.integration
+def test_regulators_cftc_cot_search(params, headers):
+    params = {p: v for p, v in params.items() if v}
+
+    query_str = get_querystring(params, [])
+    url = f"http://0.0.0.0:8000/api/v1/regulators/cftc/cot_search?{query_str}"
+    result = requests.get(url, headers=headers, timeout=10)
+    assert isinstance(result, requests.Response)
+    assert result.status_code == 200
+
+
+@pytest.mark.parametrize(
+    "params",
+    [
+        (
+            {
+                "id": "13874P",
+                "data_type": "FO",
+                "legacy_format": True,
+                "report_type": "ALL",
+                "measure": "CR",
+                "start_date": "2023-01-01",
+                "end_date": "2023-06-06",
+                "transform": "diff",
+                "provider": "nasdaq",
+            }
+        ),
+    ],
+)
+@pytest.mark.integration
+def test_regulators_cftc_cot(params, headers):
+    params = {p: v for p, v in params.items() if v}
+
+    query_str = get_querystring(params, [])
+    url = f"http://0.0.0.0:8000/api/v1/regulators/cftc/cot?{query_str}"
     result = requests.get(url, headers=headers, timeout=10)
     assert isinstance(result, requests.Response)
     assert result.status_code == 200
