+++ openbb_platform/extensions/regulators/integration/test_regulators_python.py
@@ -3,14 +3,18 @@ import pytest
 from openbb_core.app.model.obbject import OBBject
 
 
+# pylint: disable=inconsistent-return-statements
 @pytest.fixture(scope="session")
 def obb(pytestconfig):
     """Fixture to setup obb."""
 
     if pytestconfig.getoption("markexpr") != "not integration":
-        import openbb
+        import openbb  # pylint: disable=import-outside-toplevel
 
         return openbb.obb
+
+
+# pylint: disable=redefined-outer-name
 
 
 @pytest.mark.parametrize(
@@ -132,6 +136,46 @@ def test_regulators_sec_rss_litigation(params, obb):
 @pytest.mark.integration
 def test_regulators_sec_sic_search(params, obb):
     result = obb.regulators.sec.sic_search(**params)
+    assert result
+    assert isinstance(result, OBBject)
+    assert len(result.results) > 0
+
+
+@pytest.mark.parametrize(
+    "params",
+    [
+        ({"query": "grain", "provider": "nasdaq"}),
+    ],
+)
+@pytest.mark.integration
+def test_regulators_cftc_cot_search(params, obb):
+    result = obb.regulators.cftc.cot_search(**params)
+    assert result
+    assert isinstance(result, OBBject)
+    assert len(result.results) > 0
+
+
+@pytest.mark.parametrize(
+    "params",
+    [
+        (
+            {
+                "id": "13874P",
+                "data_type": "FO",
+                "legacy_format": True,
+                "report_type": "ALL",
+                "measure": "CR",
+                "start_date": "2023-01-01",
+                "end_date": "2023-06-06",
+                "transform": "diff",
+                "provider": "nasdaq",
+            }
+        ),
+    ],
+)
+@pytest.mark.integration
+def test_regulators_cftc_cot(params, obb):
+    result = obb.regulators.cftc.cot(**params)
     assert result
     assert isinstance(result, OBBject)
     assert len(result.results) > 0
