+++ openbb_platform/extensions/regulators/openbb_regulators/cftc/__init__.py
@@ -1,0 +1,1 @@
+"""Commodity Futures Trading Commission (CFTC)."""
